requirements.txt
Flask==3.0.0
python-docx==1.1.0
pybtex==0.24.0
requests==2.31.0
Jinja2==3.1.2

# 生產環境
gunicorn==21.2.0
